+++ VERSION-HISTORY.md
@@ -1,8 +1,12 @@
 # Toucan Version History & Release Notes
 
-### 1.0.3 (May 3rd, 2017)
+### [Unreleased]
 
-*  Fixed [#6](https://github.com/metabase/toucan/issues/6): Fix hydration for fields that end in `-id`. Credit: [@plexus](https://github.com/plexus)
+* Make `toucan.db/insert!` and friends use HoneySQL, so that you can use `honeysql.core/call` for SQL function calls. ([@plexus](https://github.com/plexus))
+
+### [1.0.3] (May 3rd, 2017)
+
+*  Fixed [#6](https://github.com/metabase/toucan/issues/6): Fix hydration for fields that end in `-id`. ([@plexus](https://github.com/plexus))
 
 ### 1.0.2 (Jan 27th, 2017)
 
@@ -18,3 +22,6 @@
 ### 1.0.0 (Jan 26th, 2017)
 
 *  Initial release of Toucan.
+
+[Unreleased]: https://github.com/metabase/toucan/compare/1.0.3...HEAD
+[1.0.3]: https://github.com/metabase/toucan/compare/1.0.2...1.0.3